pyproject.toml
[project]
name = "llm-rag-response-pipe"
version = "0.1.0"
description = "ZMQ-based pipeline for LLM and RAG processing with support for local and remote inference"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "click>=8.3.1",
    "pyzmq>=27.1.0",
    "requests>=2.32.5",
    "qdrant-client>=1.7.0",
    "pydantic>=2.0.0",
    "msgpack>=1.0.0",
    "sentence-transformers>=5.2.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=4.1.0",
    "pytest-mock>=3.12.0",
    "pytest-timeout>=2.2.0",
    "responses>=0.24.0",
]

[project.scripts]
llm-rag-pipe = "src.cli:cli"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = [
    "-v",
    "--strict-markers",
]
markers = [
    "unit: Unit tests for individual components",
    "integration: Integration tests for component interactions",
    "slow: Tests that take significant time to run",
    "requires_qdrant: Tests that require a running Qdrant instance",
    "requires_llm: Tests that require LLM model access",
]
timeout = 30

[tool.coverage.run]
source = ["src"]
omit = [
    "*/tests/*",
    "*/test_*.py",
    "*/__pycache__/*",
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
    "if TYPE_CHECKING:",
    "@abstractmethod",
]
